requirements.txt
grpcio-tools~=1.44
scikit-learn~=1.0.2
numpy~=1.22.3
pandas~=1.4.2